requirements.txt
# ── Core ──
fastapi==0.115.0
uvicorn==0.30.0
streamlit==1.40.0
python-dotenv==1.0.0

# ── LangChain + LangGraph ──
langchain==0.3.0
langgraph==0.2.0
langchain-core==0.3.0
langchain-community==0.3.0
langchain-anthropic==0.2.0
langchain-text-splitters==0.3.0

# ── Anthropic ──
anthropic==0.40.0

# ── Vector DB ──
qdrant-client==1.12.0

# ── Embeddings + Reranking ──
sentence-transformers==3.3.0
flashrank==0.2.9
rank-bm25==0.2.2

# ── Document Loading ──
pypdf==5.1.0

# ── HTTP ──
httpx==0.27.0

# ── Testing ──
pytest==8.3.0